Changes based on comments.
